Add ability to select photos
This is the start of adding photos to the application. It adds in the ability to take a photo from the camera and select photos from the phone.

Enough of the assign controller was also implmented to verify that we were handling the photos correctly to send them along to another view controller.

Next step is to assign the photos correctly so they have the correct details and get saved into the database.